Refactor Spiritnet runtime
